requirements.txt
kaggle
ploomber
openpyxl
pandas
mysql-connector-python
plotly
dash
dash-mantine-components
python-dotenv
